finish delete

+++ backend/controllers/goalController.js
@@ -1,10 +1,12 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User =  require('../models/userModel')
+
 //@desc Get goals
 //@route Get/api/ goals
 //@acess Private
 const getGoals =asyncHandler(async(req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json(goals)
 })
 
@@ -18,7 +20,8 @@ const setGoals = asyncHandler(async( req, res) => {
    }
 
    const goal = await Goal.create({
-    text: req.body.text
+    text: req.body.text,
+    user: req.user.id
 
    })
     res.status(200).json(goal)
@@ -35,6 +38,21 @@ const setGoals = asyncHandler(async( req, res) => {
     res.status(400)
     throw new Error('Goal not found')
     }
+
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //make sure the logged in user matches the goal user
+
+    if(goal.user.toString() !==user.id){
+        res.status(401)
+        throw new Error('user not authorized')
+    }
 //finding the goal and creating if it doesn't exist
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new:true})
 
@@ -46,13 +64,23 @@ const setGoals = asyncHandler(async( req, res) => {
 //@route DELETE/api/ goals
 //@acess Private
  const deleteGoals = asyncHandler(async( req, res) => {
+   //const goal = Goal.findByIdAndRemove(req.params.id)
+   const user = await User.findById(req.user.id)
+
+   //check for user
+   if(!user){
+       res.status(401)
+       throw new Error('User not found')
+   }
+
+   //make sure the logged in user matches the goal user
+
+   if(goal.user.toString() !==user.id){
+       res.status(401)
+       throw new Error('user not authorized')
+   }
    const goal = Goal.findByIdAndRemove(req.params.id)
-   
-    if(!goal){
-        res.status(400)
-        throw new Error('Goal not found')
-        }
-        await Goal.findByIdAndRemove(req.params.id)
+       
     res.status(200).json({id: req.params.id});
  });
